+++ R/checks.R
@@ -147,7 +147,7 @@ check_codetags <- function(path = ".", exclude_pattern = "\\.Rcheck/",
 #'           file.path(package_path, "R"))
 #' roxygen2::roxygenize(package_path)
 #' print(tarball <- get_pkg_archive_path(package_path))
-#' devtools::build(pkg = package_path, path = package_path)
+#' pkgbuild::build(pkg = package_path, path = package_path)
 #' print(check_archive(tarball))
 #' }
 check_archive <- function(path, cmdargs = NULL) {